org: Separate frontend and backend JS into separate directories
