+++ doc/scripts/index.html
@@ -43,12 +43,109 @@
 
 
     <section>
-        <article><h1>cse110-sp25-group26 Go Gitters</h1>
-<p>Project Repository for CSE110, Group 26.</p>
-<h2>Resources</h2>
+        <article><h1>BalatJack - By the Go Gitters (Team 26)</h1>
+<p>BalatJack is a lightweight, browser-based twist on Blackjack that we're building as the final project for CSE110 at UC San Diego. The focus: well-made code, a fun game, and a well-structured development process that keeps quality high despite a tight schedule.</p>
+<br>
+<h2>🛣️ Sprint Roadmap</h2>
+<table>
+<thead>
+<tr>
+<th><strong>Milestone</strong></th>
+<th><strong>What we're building</strong></th>
+</tr>
+</thead>
+<tbody>
+<tr>
+<td>May 17</td>
+<td>Core game loop, basic UI layout</td>
+</tr>
+<tr>
+<td>May 24</td>
+<td>Enhanced UI/UX, audio effects, more Jokers/Consumables</td>
+</tr>
+<tr>
+<td>May 31</td>
+<td>User testing and feedback collection, add/extend unit tests</td>
+</tr>
+<tr>
+<td>June 5</td>
+<td>Bugfixes and performance passes, documentation refresh, final deployment</td>
+</tr>
+</tbody>
+</table>
+<p>(Pulled from our <a href="/admin/misc/rules.md">Team Charter</a>)</p>
+<h2>🚀 Quick Start</h2>
+<p>Clone -&gt; Install -&gt; Test - in three commands:</p>
+<pre class="prettyprint source lang-bash"><code># 1. clone
+git clone git@github.com:cse110-sp25-group26/cse110-sp25-group26.git
+cd cse110-sp25-group26
+
+# 2. install deps
+npm install
+
+# 3. sanity check
+npm run test
+</code></pre>
+<p>Need more detail? Read the <a href="/doc/GETTINGSTARTED.md">Getting Started Guide</a>.</p>
+<h2>📁 Repository Layout</h2>
+<pre class="prettyprint source"><code>.
+├── admin/               # Class-related documentation
+├── source/              # Game code & assets
+│   ├── scripts/         # JavaScript classes & helpers
+│   ├── css/             # Component‑scoped CSS
+│   └── res/             # Images, audio, etc.
+├── doc/                 # Project docs actively used for writing code
+├── specs/               # Architectural and design records
+└── index.html           # Entry point – game start for browser
+</code></pre>
+<p>Full spec: <a href="/doc/STRUCTURE.md">doc/STRUCTURE.md</a></p>
+<h2>🛠️ Tech Stack</h2>
 <ul>
-<li><a href="/admin/team.md">Team Page</a>: Meet the Team behind this project</li>
-</ul></article>
+<li><strong>HTML + CSS</strong> - semantic, responsive visual pages</li>
+<li><strong>Vanilla JS</strong> - No frameworks!</li>
+<li><strong>JSDoc</strong> - API docs, compiles to <code>doc/scripts/</code> on <code>npm run docs</code></li>
+<li><strong>Jest</strong> - Unit tests (<code>npm run test</code>)</li>
+<li><strong>ESLint</strong> - Code linting (<code>npx eslint .</code>)</li>
+</ul>
+<h2>🧑‍💻 Contributing</h2>
+<ol>
+<li>Branch naming &amp; rules -&gt; <a href="/doc/BRANCHING.md">doc/BRANCHING.md</a></li>
+<li>Style -&gt; <a href="/doc/CODE_STYLE.md">doc/CODE_STYLE.md</a></li>
+<li>Tests libe in <code>source/scripts/tests/</code></li>
+<li>Open a PR against <code>develop</code>; follow the <a href="/doc/REVIEWING.md">Review Checklist</a></li>
+</ol>
+<h2>🏗️ Current CI Pipeline</h2>
+<table>
+<thead>
+<tr>
+<th>Step</th>
+<th>Command</th>
+<th>Automated Trigger</th>
+</tr>
+</thead>
+<tbody>
+<tr>
+<td>Tests</td>
+<td><code>npm run test</code></td>
+<td>PRs -&gt; <code>develop</code>/<code>main</code></td>
+</tr>
+<tr>
+<td>Docs</td>
+<td><code>npm run docs</code></td>
+<td>PRs -&gt; <code>develop</code>/<code>main</code></td>
+</tr>
+<tr>
+<td>Lint</td>
+<td><code>npx eslint .</code></td>
+<td>PRs -&gt;<code> develop</code>/<code>main</code></td>
+</tr>
+</tbody>
+</table>
+<p>All checks must pass before a protected branch can merge.</p>
+<h2>👥 Team &amp; Credits</h2>
+<p>Faces &amp; roles: <a href="/admin/team.md">Team Page</a> - Charter: <a href="/admin/misc/rules.md">admin/misc/rules.md</a></p>
+<h2>📝 License</h2>
+<p>TBD - must determine before making repository public</p></article>
     </section>
